Added explicit casts as appropriate to avoid compiler warnings (MS VC++ 7.0)

+++ Utilities/aafviewer/src/aaf2dot/DotRecordNode.cpp
@@ -68,9 +68,9 @@ DotRecordNodeAttribute::GetWidth( InstanceMapperProfile &profile )
       if ( newPos == -1 )
       {
 	 done = true;
-	 if ( ( ltdAttributeString.size() - pos ) > maxWidth )
+	 if ( ( (int)ltdAttributeString.size() - pos ) > maxWidth )
 	 {
-	    maxWidth = ltdAttributeString.size() - pos;
+	    maxWidth = (int)ltdAttributeString.size() - pos;
 	 }
       }
       else
@@ -120,8 +120,8 @@ DotRecordNode::Write( ofstream &dotFile, InstanceMapperProfile &profile )
       fontSize = atoi( fontSizeString.c_str() );
    }
 
-   float fontSizeFactor = 0.013 * fontSize;
-   float nodeWidth = maxWidth * fontSizeFactor;
+   double fontSizeFactor = 0.013 * fontSize;
+   double nodeWidth = maxWidth * fontSizeFactor;
 
    ostringstream widthString;
    widthString << nodeWidth;
